Fix infinite loop reported by Jakup Wilk via Niels Thykier.

+++ t1lib.c
@@ -244,9 +244,9 @@ void
 process_pfb(FILE *ifp, const char *ifp_filename, struct font_reader *fr)
 {
   int blocktyp = 0;
-  int block_len = 0;
+  unsigned block_len = 0;
   int c = 0;
-  int filepos = 0;
+  unsigned filepos = 0;
   int linepos = 0;
   char line[LINESIZE];
 
@@ -260,7 +260,7 @@ process_pfb(FILE *ifp, const char *ifp_filename, struct font_reader *fr)
 	if (c == EOF || blocktyp == EOF)
 	  error("%s corrupted: no end-of-file marker", ifp_filename);
 	else
-	  error("%s corrupted: bad block marker at position %d",
+	  error("%s corrupted: bad block marker at position %u",
 		ifp_filename, filepos);
 	blocktyp = PFB_DONE;
       }
@@ -270,9 +270,9 @@ process_pfb(FILE *ifp, const char *ifp_filename, struct font_reader *fr)
       block_len = getc(ifp) & 0xFF;
       block_len |= (getc(ifp) & 0xFF) << 8;
       block_len |= (getc(ifp) & 0xFF) << 16;
-      block_len |= (getc(ifp) & 0xFF) << 24;
+      block_len |= (unsigned) (getc(ifp) & 0xFF) << 24;
       if (feof(ifp)) {
-	error("%s corrupted: bad block length at position %d",
+	error("%s corrupted: bad block length at position %u",
 	      ifp_filename, filepos);
 	blocktyp = PFB_DONE;
 	goto done;
@@ -282,11 +282,11 @@ process_pfb(FILE *ifp, const char *ifp_filename, struct font_reader *fr)
 
     /* read the block in its entirety, in LINESIZE chunks */
     while (block_len > 0) {
-      int rest = LINESIZE - 1 - linepos; /* leave space for '\0' */
-      int n = (block_len > rest ? rest : block_len);
+      unsigned rest = LINESIZE - 1 - linepos; /* leave space for '\0' */
+      unsigned n = (block_len > rest ? rest : block_len);
       int actual = fread(line + linepos, 1, n, ifp);
-      if (actual != n) {
-	error("%s corrupted: block short by %d bytes at position %d",
+      if (actual != (int) n) {
+	error("%s corrupted: block short by %u bytes at position %u",
 	      ifp_filename, block_len - actual, filepos);
 	block_len = actual;
       }
@@ -311,7 +311,7 @@ process_pfb(FILE *ifp, const char *ifp_filename, struct font_reader *fr)
  done:
   c = getc(ifp);
   if (c != EOF)
-    error("%s corrupted: data after PFB end marker at position %d",
+    error("%s corrupted: data after PFB end marker at position %u",
 	  ifp_filename, filepos - 2);
   fr->output_end();
 }
